moved viz.tsx code inside of index.tsx and removed unnecessary code
